Fix Deployment
NextAuth.js is so out of date at this point it's getting depressing,

This commit is a fix for the Next.js requirement that Route Handlers must only export functions named after HTTP methods and that take in  a Request object and return Response object. For NextAuth to work and to be able to import authOptions elsewhere (for use in `getSession()` for example), it was moved to `/lib/auth.ts`, then imported to `/app/api/auth/[...auth].ts`. This eliminates the need to export non-HTTP methods functions from a Route Handler.

Seriously NextAuth, please update your shit.